projects/shared_lib/pyproject.toml
[project]
name = "shared-lib"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "google-generativeai>=0.8",
    "gtts>=2.5",
    "requests>=2.31",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
